refactor(json-crdt-patch): 💡 rename CRDT writer and reader file names
