+++ src/json-crdt-patch/util/binary/CrdtReader.ts
