+++ src/json-crdt-patch/util/binary/CrdtWriter.ts
