+++ src/json-crdt-patch/util/binary/__tests__/CrdtWriter.vuint57.spec.ts
@@ -1,9 +1,9 @@
-import {CrdtWriter} from '../CrdtEncoder';
+import {CrdtWriter} from '../CrdtWriter';
 
 const encoder = new CrdtWriter(1);
 const encode = (num: number): Uint8Array => {
   encoder.reset();
-  encoder.vu39(num);
+  encoder.vu57(num);
   return encoder.flush();
 };
 
@@ -56,14 +56,31 @@ test('encodes five byte integers', () => {
   );
 });
 
-test('encodes sixth byte up to 4 bits', () => {
+test('encodes six byte integers', () => {
   expect(encode(0b1_1111111_1111111_1111111_1111111_1111111)).toEqual(
     new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_0000001]),
   );
   expect(encode(0b1111101_1111111_0000000_1111111_1111111_1111111)).toEqual(
-    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_0000000, 0b1_1111111, 0b0_0001101]),
+    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_0000000, 0b1_1111111, 0b0_1111101]),
   );
   expect(encode(0b1111111_1111111_1111111_1111111_1111111_1111111)).toEqual(
-    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_0001111]),
+    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_1111111]),
+  );
+});
+
+test('encodes seven byte integers', () => {
+  expect(encode(0b1_1111111_1111111_1111111_1111111_1111111_1111111)).toEqual(
+    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_0000001]),
+  );
+  expect(encode(0b1111111_1111111_1111111_1111111_1111111_1111111_1111111)).toEqual(
+    new Uint8Array([0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_1111111]),
+  );
+});
+
+test('encodes eight byte integers', () => {
+  expect(encode(0b1_1111111_1111111_1111111_1111111_1111111_1111111_1111111)).toEqual(
+    new Uint8Array([
+      0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b1_1111111, 0b0_0000001,
+    ]),
   );
 });
